+++ src/engine/index.ts
@@ -1,5 +1,6 @@
 export * from "./domain";
 export * from "./generation";
+export * from "./news";
 export * from "./rng";
 export * from "./simulation";
 export * from "./transfers";